Integration of Foundation Framework into Contentia Ipsumio
Integration of Framework Studies Begins with this Upload ... still not
certain how github version control works but this is what I am looking
for ... i just need to know how spin the record forward and backwords on
this GUI needs that slider control to manage my workflow but for now I
am competent at a entry level position with GitHub.

If I could cut between uploads like frames I could keyframe a
performance is the DreamWeaver Timeline Function has not be depricated.

There was a time when websites had Timelines ... if you used
Dreamweaver.

Nicholas